+++ src/components/TypeaheadSelect/TypeaheadSelect.tsx
@@ -4,6 +4,7 @@ import {
   ComboboxOptions,
   Label,
 } from '@headlessui/react';
+import { AnchorProps } from '@headlessui/react/dist/internal/floating';
 import classNames from 'classnames';
 import {
   Fragment,
@@ -35,6 +36,7 @@ export interface TypeaheadSelectProps<
   Values extends FieldValues,
 > extends UseTypeaheadQueryOptions<DataItem>,
     FormFieldProps<Values> {
+  anchor?: AnchorProps;
   badgeColor?: BadgeColor;
   bordered?: boolean;
   color?: InputColor;
@@ -52,6 +54,7 @@ export interface TypeaheadSelectProps<
   menuSize?: MenuSize;
   multi?: true;
   onKeyDown?: KeyboardEventHandler<HTMLInputElement>;
+  portal?: boolean;
   size?: InputSize;
 }
 
@@ -59,6 +62,7 @@ export const TypeaheadSelect = <
   DataItem extends GenericDataType,
   Values extends FieldValues,
 >({
+  anchor,
   badgeColor,
   bordered,
   className,
@@ -84,6 +88,7 @@ export const TypeaheadSelect = <
   onKeyDown,
   options: optionsArray,
   placeholder,
+  portal,
   showDirty,
   size,
 }: TypeaheadSelectProps<DataItem, Values>) => {
@@ -127,7 +132,7 @@ export const TypeaheadSelect = <
   const Component = multi === true ? ComboboxMulti : ComboboxSingle;
   return (
     <Component
-      className={classNames('relative', className)}
+      className={className}
       disabled={disabled}
       name={name}
       selectedItems={selectedItems}
@@ -207,11 +212,10 @@ export const TypeaheadSelect = <
       </fieldset>
       {showDropdown ? (
         <ComboboxOptions
+          anchor={anchor}
           as={Menu}
-          className={classNames(
-            'absolute rounded-box p-2 shadow-xl',
-            menuClassName,
-          )}
+          className={classNames('rounded-box p-2 shadow-lg', menuClassName)}
+          portal={portal}
           ref={dropdownRef}
           size={menuSize}
           static